Migrating MORE commands
Migrating more Kord commands to Discord InteraKTions Unleashed.
