--- loritta-bot-discord/src/main/kotlin/net/perfectdreams/loritta/cinnamon/discord/interactions/vanilla/fun/declarations/SummonCommand.kt
@@ -1,31 +1,0 @@
-package net.perfectdreams.loritta.cinnamon.discord.interactions.vanilla.`fun`.declarations
-
-import net.perfectdreams.loritta.common.locale.LanguageManager
-import net.perfectdreams.loritta.common.utils.TodoFixThisData
-import net.perfectdreams.loritta.i18n.I18nKeysData
-import net.perfectdreams.loritta.morenitta.LorittaBot
-import net.perfectdreams.loritta.common.commands.CommandCategory
-import net.perfectdreams.loritta.cinnamon.discord.interactions.commands.CinnamonSlashCommandDeclarationWrapper
-import net.perfectdreams.loritta.cinnamon.discord.interactions.vanilla.`fun`.BemBoladaExecutor
-import net.perfectdreams.loritta.cinnamon.discord.interactions.vanilla.`fun`.FaustaoExecutor
-import net.perfectdreams.loritta.cinnamon.discord.interactions.vanilla.`fun`.TioDoPaveExecutor
-
-class SummonCommand(languageManager: LanguageManager) : CinnamonSlashCommandDeclarationWrapper(languageManager) {
-    companion object {
-        val I18N_PREFIX = I18nKeysData.Commands.Command.Summon
-    }
-
-    override fun declaration() = slashCommand(I18N_PREFIX.Label, CommandCategory.FUN, TodoFixThisData) {
-        subcommand(I18N_PREFIX.Tiodopave.Label, I18N_PREFIX.Tiodopave.Description) {
-            executor = { TioDoPaveExecutor(it) }
-        }
-
-        subcommand(I18N_PREFIX.Faustao.Label, I18N_PREFIX.Faustao.Description) {
-            executor = { FaustaoExecutor(it) }
-        }
-
-        subcommand(I18N_PREFIX.Kenji.Label, I18N_PREFIX.Kenji.Description) {
-            executor = { BemBoladaExecutor(it) }
-        }
-    }
-}
