--- loritta-bot-discord/src/main/kotlin/net/perfectdreams/loritta/cinnamon/discord/interactions/vanilla/fun/texttransform/TextClapExecutor.kt
@@ -1,29 +1,0 @@
-package net.perfectdreams.loritta.cinnamon.discord.interactions.vanilla.`fun`.texttransform
-
-import net.perfectdreams.discordinteraktions.common.commands.options.SlashCommandArguments
-import net.perfectdreams.loritta.morenitta.LorittaBot
-import net.perfectdreams.loritta.cinnamon.discord.interactions.cleanUpForOutput
-import net.perfectdreams.loritta.cinnamon.discord.interactions.commands.ApplicationCommandContext
-import net.perfectdreams.loritta.cinnamon.discord.interactions.commands.CinnamonSlashCommandExecutor
-import net.perfectdreams.loritta.cinnamon.discord.interactions.commands.options.LocalizedApplicationCommandOptions
-import net.perfectdreams.loritta.cinnamon.discord.interactions.vanilla.`fun`.declarations.TextTransformCommand
-
-class TextClapExecutor(loritta: LorittaBot) : CinnamonSlashCommandExecutor(loritta) {
-    inner class Options : LocalizedApplicationCommandOptions(loritta) {
-        val text = string("text", TextTransformCommand.I18N_PREFIX.Clap.Options.Text(TextTransformCommand.CLAP_EMOJI))
-
-        val emoji = optionalString("emoji", TextTransformCommand.I18N_PREFIX.Clap.Options.Emoji)
-    }
-
-    override val options = Options()
-
-    override suspend fun execute(context: ApplicationCommandContext, args: SlashCommandArguments) {
-        val text = cleanUpForOutput(context, args[options.text])
-        val emoji = args[options.emoji] ?: TextTransformCommand.CLAP_EMOJI
-
-        context.sendReply(
-            content = "$emoji${text.split(" ").joinToString(emoji)}$emoji",
-            prefix = "✍"
-        )
-    }
-}
