--- loritta-bot-discord/src/main/kotlin/net/perfectdreams/loritta/cinnamon/discord/interactions/vanilla/fun/texttransform/TextMockExecutor.kt
@@ -1,33 +1,0 @@
-package net.perfectdreams.loritta.cinnamon.discord.interactions.vanilla.`fun`.texttransform
-
-import net.perfectdreams.discordinteraktions.common.commands.options.SlashCommandArguments
-import net.perfectdreams.loritta.morenitta.LorittaBot
-import net.perfectdreams.loritta.cinnamon.discord.interactions.cleanUpForOutput
-import net.perfectdreams.loritta.cinnamon.discord.interactions.commands.ApplicationCommandContext
-import net.perfectdreams.loritta.cinnamon.discord.interactions.commands.CinnamonSlashCommandExecutor
-import net.perfectdreams.loritta.cinnamon.discord.interactions.commands.options.LocalizedApplicationCommandOptions
-import net.perfectdreams.loritta.cinnamon.discord.interactions.vanilla.`fun`.declarations.TextTransformCommand
-import kotlin.random.Random
-
-class TextMockExecutor(loritta: LorittaBot) : CinnamonSlashCommandExecutor(loritta) {
-    inner class Options : LocalizedApplicationCommandOptions(loritta) {
-        val text = string("text", TextTransformCommand.I18N_PREFIX.Mock.Description)
-    }
-
-    override val options = Options()
-
-    override suspend fun execute(context: ApplicationCommandContext, args: SlashCommandArguments) {
-        val text = cleanUpForOutput(context, args[options.text])
-
-        // Create a random based on the text's hash code
-        val random = Random(text.hashCode())
-
-        val mockedText = text.mapIndexed { index, c -> if (random.nextBoolean()) c.uppercaseChar() else c.lowercaseChar() }
-            .joinToString("")
-
-        context.sendReply(
-            content = mockedText,
-            prefix = "✍"
-        )
-    }
-}
