+++ loritta-bot-discord/src/main/kotlin/net/perfectdreams/loritta/cinnamon/discord/voice/LorittaVoiceConnection.kt
@@ -14,7 +14,7 @@ import kotlin.time.Duration.Companion.minutes
 
 data class LorittaVoiceConnection(
     private val guild: Guild,
-    var channelId: Snowflake, // Users can move Loritta to another channel
+    var channelId: Long, // Users can move Loritta to another channel
     private val audioManager: AudioManager,
     private val audioProvider: LorittaAudioProvider,
     private val audioClipProviderNotificationChannel: Channel<Unit>
@@ -34,7 +34,7 @@ data class LorittaVoiceConnection(
         audioClips.send(audioClip)
     }
 
-    suspend fun switchChannel(channelId: Snowflake) {
+    suspend fun switchChannel(channelId: Long) {
         if (this.channelId != channelId) {
             this.channelId = channelId
             audioManager.openAudioConnection(guild.getVoiceChannelById(channelId.toLong()))
@@ -77,6 +77,6 @@ data class LorittaVoiceConnection(
 
     class AudioClipInfo(
         val frames: List<ByteArray>,
-        val channelId: Snowflake
+        val channelId: Long
     )
 }
