--- loritta-bot-discord/src/main/kotlin/net/perfectdreams/loritta/morenitta/commands/vanilla/fun/VaporQualidadeCommand.kt
@@ -1,34 +1,0 @@
-package net.perfectdreams.loritta.morenitta.commands.vanilla.`fun`
-
-import net.perfectdreams.loritta.morenitta.commands.AbstractCommand
-import net.perfectdreams.loritta.morenitta.commands.CommandContext
-import net.perfectdreams.loritta.morenitta.utils.escapeMentions
-import net.perfectdreams.loritta.morenitta.messages.LorittaReply
-import net.perfectdreams.loritta.common.locale.BaseLocale
-import net.perfectdreams.loritta.common.locale.LocaleKeyData
-import net.perfectdreams.loritta.common.utils.text.VaporwaveUtils
-import net.perfectdreams.loritta.morenitta.utils.OutdatedCommandUtils
-import net.perfectdreams.loritta.morenitta.LorittaBot
-
-class VaporQualidadeCommand(loritta: LorittaBot) : AbstractCommand(loritta, "vaporqualidade", category = net.perfectdreams.loritta.common.commands.CommandCategory.FUN) {
-	override fun getDescriptionKey() = LocaleKeyData("commands.command.vaporquality.description")
-	override fun getExamplesKey() = LocaleKeyData("commands.command.vaporquality.examples")
-
-	// TODO: Fix Usage
-	// TODO: Fix Detailed Usage
-
-	override suspend fun run(context: CommandContext,locale: BaseLocale) {
-		OutdatedCommandUtils.sendOutdatedCommandMessage(context, locale, "text vaporquality")
-
-		if (context.args.isNotEmpty()) {
-			val qualidade = VaporwaveUtils.vaporwave(context.args.joinToString(" ").toCharArray().joinToString(" ")).toUpperCase()
-					.escapeMentions()
-
-			context.reply(
-                    LorittaReply(message = qualidade, prefix = "✍")
-			)
-		} else {
-			this.explain(context)
-		}
-	}
-}
